gmail_mcp/mcp/server.py
"""
MCP Server Module

This module provides the core MCP server implementation, including route setup and request handling.
"""

import logging
from typing import Dict, Any, List, Optional

from fastapi import FastAPI, Depends, HTTPException, Request
from fastapi.responses import JSONResponse
from pydantic import BaseModel

from gmail_mcp.utils.logger import get_logger
from gmail_mcp.utils.config import get_config
from gmail_mcp.mcp.schemas import (
    MessageRequest,
    MessageResponse,
    ActionRequest,
    ActionResponse,
    SearchRequest,
    SearchResponse,
)

# Get logger
logger = get_logger(__name__)


class VersionResponse(BaseModel):
    """MCP version response model."""
    
    version: str


def setup_mcp_routes(app: FastAPI) -> None:
    """
    Set up MCP routes on the FastAPI application.
    
    Args:
        app (FastAPI): The FastAPI application.
    """
    # MCP version endpoint
    @app.get("/.well-known/mcp-version")
    async def mcp_version() -> VersionResponse:
        """Get the MCP version supported by this server."""
        config = get_config()
        return VersionResponse(version=config["mcp_version"])
    
    # MCP messages endpoint
    @app.post("/messages")
    async def messages(request: MessageRequest) -> MessageResponse:
        """
        Process a message request from Claude.
        
        Args:
            request (MessageRequest): The message request.
            
        Returns:
            MessageResponse: The message response.
        """
        logger.info(f"Received message request: {request.model_dump_json()}")
        
        # TODO: Implement message processing
        
        # For now, return a placeholder response
        return MessageResponse(
            response_id="placeholder_response_id",
            content="This is a placeholder response. The Gmail MCP server is still under development.",
            context=[],
            actions=[],
        )
    
    # MCP actions endpoint
    @app.post("/actions")
    async def actions(request: ActionRequest) -> ActionResponse:
        """
        Process an action request from Claude.
        
        Args:
            request (ActionRequest): The action request.
            
        Returns:
            ActionResponse: The action response.
        """
        logger.info(f"Received action request: {request.model_dump_json()}")
        
        # TODO: Implement action processing
        
        # For now, return a placeholder response
        return ActionResponse(
            response_id="placeholder_response_id",
            content="This is a placeholder response. The action functionality is still under development.",
            context=[],
        )
    
    # MCP search endpoint
    @app.post("/search")
    async def search(request: SearchRequest) -> SearchResponse:
        """
        Process a search request from Claude.
        
        Args:
            request (SearchRequest): The search request.
            
        Returns:
            SearchResponse: The search response.
        """
        logger.info(f"Received search request: {request.model_dump_json()}")
        
        # TODO: Implement search functionality
        
        # For now, return a placeholder response
        return SearchResponse(
            response_id="placeholder_response_id",
            content="This is a placeholder response. The search functionality is still under development.",
            context=[],
        ) 