pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gmail-mcp"
version = "0.1.0"
description = "A Model Context Protocol (MCP) server for Gmail integration with Claude Desktop"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "mcp-python-sdk>=0.1.0",
    "uvicorn>=0.23.2",
    "pydantic>=2.4.2",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "google-auth>=2.23.3",
    "google-auth-oauthlib>=1.1.0",
    "google-api-python-client>=2.105.0",
    "cryptography>=41.0.4",
    "python-jose>=3.3.0",
    "spacy>=3.7.2",
    "nltk>=3.8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.2",
    "pytest-cov>=4.1.0",
    "black>=23.9.1",
    "isort>=5.12.0",
    "mypy>=1.6.1",
    "ruff>=0.0.292",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/gmail-mcp"
"Bug Tracker" = "https://github.com/yourusername/gmail-mcp/issues"

[tool.setuptools]
packages = ["gmail_mcp"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "B", "I"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*" 